tool_convert/convert_csv_2/backend/requirements.txt
Django==4.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pandas==1.4.2
celery==5.2.3
redis==4.3.4